smart-receipt-tracker/function/requirements.txt
azure-functions
azure-storage-blob
azure-ai-formrecognizer
azure-core
